==> scripts/12_Structure.Rmd <==
---
title: "Genetic Structure Results"
author: "C. E. Moore"
output: 
  pdf_document:
    latex_engine: pdflatex
urlcolor: blue
header-includes:
  \usepackage{helvet}
  \renewcommand{\familydefault}{\sfdefault}
---
Last updated on `r Sys.Date()`  
  
*Note: Microsat data previously formatted from geneious output using 01a_DataPrep.R and then filtered for missing data in 10_MicrosatSummary.Rmd*  

**Contents:**  
1. F~ST~ (hierfstat)  
2. F~ST~ (pegas)  
3. Pairwise F~ST~  
4. Dps (shared alleles)  
5. G~ST~
6. PCA of pops  
7. DAPCs  
  
```{r paths and data load, echo=FALSE, message=FALSE}
## packages
library(adegenet); library(ape); library(pegas); library(knitr); library(tidyverse); library(hierfstat); library(PopGenReport); library(ggpubr)

knitr::opts_chunk$set(echo = FALSE)

## home path
PATH <- "/home/chloe9mo/Documents/Projects/temporal_aztreefrog"

## load and edit data
source(file = paste0(PATH, "/code/01c_DataLoad.R"))
```
**Loaded packages:** `r (.packages())`  

## Global and Locus F Statistics  
### hierfstat F~ST~  
``` {r hierfstat fst, message=FALSE, results='asis'}
g.fst <- data.frame(unlist(ind.names), stringsAsFactors = F) %>% 
  add_column(Fst = NA, Fis = NA) %>% rename(group = unlist.ind.names.)

for (i in 1:length(ind.list)) {
  
  g.fst[g.fst$group == ind.names[[i]],]$Fst <- wc(ind.list[[i]])$FST
  g.fst[g.fst$group == ind.names[[i]],]$Fis <- wc(ind.list[[i]])$FIS
  
}

kable(g.fst, caption = "Global F-statistics from hierfstat")

g.fst.ci <- data.frame(unlist(ind.names), stringsAsFactors = F) %>% 
  add_column(Fst = NA, CI_ll = NA, CI_ul = NA) %>% rename(group = unlist.ind.names.)

for (i in 1:length(ind.list)) {
  
  ci <- genind2hierfstat(ind.list[[i]])
  print(kable(boot.vc(ci[1], ci[-1])$ci, caption = paste0("Confidence Interval for ", ind.names[[i]], " F-stats")))
  bci <- boot.vc(ci[1], ci[-1], nboot = 10000)
  g.fst.ci[g.fst.ci$group == ind.names[[i]],]$Fst <- bci$ci[2,2]
  g.fst.ci[g.fst.ci$group == ind.names[[i]],]$CI_ll <- bci$ci[1,2]
  g.fst.ci[g.fst.ci$group == ind.names[[i]],]$CI_ul <- bci$ci[3,2]
  sig.fst <- boot.ppfst(ci, nboot = 10000)
  # print(sig.fst$ll)
  cat("\n",ifelse(any(sig.fst$ll<0, na.rm = T), "At least one pairwise Fst not sig. diff from 0", "All pairwise Fst sig. diff from 0"),"\n\n")
  
}

g.fst.ci %>% 
  left_join(., g.fst %>% select(group, Fis)) %>%
  mutate(group = case_when(group == "All Yrs - Indiv x Pop" ~ "All_Yrs",
                           group == "All Yrs - Indiv x Year" ~ "All_Pops",
                           group == "Indiv 2014" ~ "2014",
                           group == "Indiv 2019" ~ "2019",
                           group == "Indiv 2021" ~ "2021")) %>%
  write.csv(., file=paste0(PATH, "/results_tables/Fst_global_hierfstat.csv"), row.names = F)
```

``` {r hierfstat global fst plot setup, message=F, eval=F}
g.fst.ci <- g.fst.ci %>% mutate(gr = case_when(group %in% c("All Yrs - Indiv x Pop", "All Yrs - Indiv x Year") ~ "all",
                                               group %in% c("Indiv 2014", "Indiv 2019", "Indiv 2021") ~ "one"))
g.fst.ci.lin <- g.fst.ci %>%
  mutate(across(.cols = c('Fst', 'CI_ll', 'CI_ul'), .fns = function(x) (x / (1 - x)), .names = "{.col}_lin"))

pf <- ggplot(data = g.fst.ci, aes(x=group, y=Fst)) +
  geom_rect(xmin = 0, xmax = 2.5, ymin = -Inf, ymax = Inf, fill="#E5E4E2", alpha = 0.1) +
  geom_errorbar(aes(x=group, ymin=CI_ll, ymax=CI_ul), color="black", width=0.3) +
  geom_line(color="black") +
  geom_point(aes(fill=gr), shape=21, color="black", size=6) +
  # scale_fill_manual(values = c("#FFF2EB", "#CC4E00")) + #oranges
  scale_fill_manual(values = c("#2B847C", "#DF9F1F")) + #match structure plot color schemes
  scale_x_discrete(labels=c("All x Pop", "All x Year", "2014", "2019", "2021")) +
  geom_vline(aes(xintercept=2.5), linetype="dashed") +
  labs(y=bquote('Global' ~F[ST])) +
  #ggtitle(bquote('Global' ~F[ST]~ 'with 95% confidence intervals')) +
  theme(legend.position = "none",
        panel.background = element_blank(),
        panel.grid = element_line(color="#E5E4E2"),
        title = element_text(size=17),
        axis.title.x = element_blank(),
        axis.title.y = element_text(size = 15),
        axis.text.x = element_text(size = 15),
        axis.text.y = element_text(size = 12))
# pf

ggsave(filename = paste0(PATH, "/figures/global_fst_points.png"), plot = pf, width = 6, height = 6)
```
\begin{figure}
\centering
\includegraphics[width=0.5\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/global_fst_points.png"}
\caption{Comparison of global Fst values with single year and multi-year groupings, population structure vs year structure.}
\end{figure}

---

\pagebreak
### pegas F~ST~
``` {r pegas fst}
ftab <- lapply(ind.list, function(x) pegas::Fst(as.loci(x)))
lftab <- lapply(ind.list, function(x) as.data.frame(pegas::Fst(as.loci(x))) %>% rownames_to_column(var="locus"))

l.fst <- lftab[[1]] %>% 
  left_join(., lftab[[2]], by="locus") %>% 
  left_join(., lftab[[3]], by="locus") %>% 
  left_join(., lftab[[4]], by="locus") %>%
  left_join(., lftab[[5]], by="locus") %>%
  mutate(across(where(is.numeric), round, 3))

l.fst.fst <- l.fst %>% select(locus, contains("Fst"))
names(l.fst.fst) <- c("locus", unlist(ind.names))

l.fst.fis <- l.fst %>% select(locus, contains("Fis"))
names(l.fst.fis) <- c("locus", unlist(ind.names))

l.fst.fit <- l.fst %>% select(locus, contains("Fit"))
names(l.fst.fit) <- c("locus", unlist(ind.names))

kable(l.fst.fst, caption="Per-Locus Fst", booktabs=T, valign= 't')
kable(l.fst.fis, caption="Per-Locus Fis", booktabs=T, valign= 't')
kable(l.fst.fit, caption="Per-Locus Fit", booktabs=T, valign= 't')
  
g.fst <- data.frame(unlist(ind.names), stringsAsFactors = F) %>% 
  add_column(Fst = NA, Fis = NA, Fit = NA) %>% rename(group = unlist.ind.names.)
for (i in 1:length(ftab)) {
  
  g.fst[g.fst$group == ind.names[[i]],]$Fst <- as.data.frame(t(colMeans(ftab[[i]])))$Fst
  g.fst[g.fst$group == ind.names[[i]],]$Fit <- as.data.frame(t(colMeans(ftab[[i]])))$Fit
  g.fst[g.fst$group == ind.names[[i]],]$Fis <- as.data.frame(t(colMeans(ftab[[i]])))$Fis
  
}
kable(g.fst, caption = "Global F-statistics from pegas")
```

---

\bigskip
### Pairwise F~ST~  
*note: \textcolor{green}{Nei's} method used*  
``` {r add year_pop, message=F}
#redo the yr grouping genind
yr.coord <- coord %>% dplyr::select(year_pop, lon, lat) %>% unique() %>% remove_rownames() %>% column_to_rownames("year_pop")

ind_aztf_yr <- df2genind(aztf[,c(6:22)], sep = "/", ind.names = aztf$ID, NA.char = "-1/-1")
ind_aztf_yr@pop <- as.factor(aztf$year_pop)
ind_aztf_yr$other$xy <- subset(yr.coord, rownames(yr.coord) %in% aztf$year_pop) #assign coordinates

ind.list[[6]] <- ind_aztf_yr
pop.list[[6]] <- genind2genpop(ind_aztf_yr, quiet = T)
ind.names[[6]] <- "Year_Pop"
pop.names[[6]] <- "Year_Pop"

rm(yr.coord, ind_aztf_yr)
```
``` {r pairwise fst heatmaps, message=F, eval=F}
#make pairwise heat plot
heat.l <- vector(mode="list", length=length(ind.list))
for (i in 1:length(ind.list)) {
  #get pfst values
  pw <- hierfstat::genet.dist(ind.list[[i]], method = "Nei87")
  pw <- as.data.frame(as.matrix(pw))
  pw[upper.tri(pw)] <- NA
  
  #make ggplot friendly
  pw.long <- pw %>%
    rownames_to_column(var = "pop") %>%
    pivot_longer(cols = -pop, names_to = "pop2", values_to = "pFst") %>%
    mutate(pop = as.factor(as.numeric(pop)),
           pop2 = as.factor(as.numeric(pop2)))
  
  #heatmap it
  heat.l[[i]] <- ggplot() + 
    geom_tile(data=pw.long, aes(pop, pop2, fill=pFst)) +
    scale_fill_viridis_c(na.value = NA) +
    labs(title=ind.names[[i]], subtitle = paste0("Range is ", min(pw[pw>0], na.rm=T), "-", max(pw, na.rm = T)),
         y = "pop") +
    theme_bw()
  ggsave(paste0(PATH, "/figures/pairwiseFst_heatmap_", ind.names[[i]], ".png"), plot = heat.l[[i]])
  write.csv(pw, file = paste0(PATH, "/results_tables/", gsub(" ", "_", ind.names[[i]]), "_pairwiseFst.csv"), row.names = T)
}

#getting temporal pairwise fst table with significance
g.yp <- genind2hierfstat(ind.list[[6]])
yp.fst <- boot.ppfst(g.yp, nboot = 10000)
#lower ci
ll <- yp.fst$ll %>%
  as.data.frame() %>%
  rownames_to_column(., "X2") %>%
  pivot_longer(cols = starts_with("20"), names_to = "X", values_to = "pFst_ll") %>%
  filter(!is.na(pFst_ll)) %>%
  filter(X != X2)
#upper ci
ul <- yp.fst$ul %>%
  as.data.frame() %>%
  rownames_to_column(., "X2") %>%
  pivot_longer(cols = starts_with("20"), names_to = "X", values_to = "pFst_ul") %>%
  filter(!is.na(pFst_ul)) %>%
  filter(X != X2)
#make it long
pw <- hierfstat::genet.dist(ind.list[[6]], method = "Nei87")
pw <- as.data.frame(as.matrix(pw)) %>%
  relocate('2014_1', '2021_1',
           '2014_3', '2021_3',
           '2014_2', '2019_2', '2021_2',
           '2014_4', '2019_4', '2021_4',
           '2014_5', '2019_5', '2021_5',
           '2014_6', '2019_6', '2021_6',
           '2014_7', '2019_7', '2021_7',
           '2014_8', '2021_8',
           '2019_9', '2021_9')
pw <- pw %>%
  rownames_to_column(var="group") %>%
  arrange(factor(group, levels = names(pw))) %>%
  column_to_rownames("group")
write.csv(pw, file = paste0(PATH, "/results_tables/", gsub(" ", "_", ind.names[[6]]), "_pairwiseFst.csv"), row.names = T)

pw <- read.csv(paste0(PATH, "/results_tables/Year_Pop_pairwiseFst.csv"))
pw.long <- pw %>%
  pivot_longer(cols = starts_with("X2"), names_to = "X2", values_to = "pFst") %>%
  filter(!is.na(pFst)) %>%
  mutate(X2 = gsub("X", "", X2)) %>%
  filter(X != X2) %>%
  left_join(., ul) %>%
  left_join(., ll)
write.csv(pw.long, file=paste0(PATH, "/results_tables/Year_Pop_Long_pFst.csv"), row.names = F)

pw.long <- pw.long %>% 
  mutate(X = fct_relevel(factor(X), gsub("X", "", names(pw[,-1]))),
         X2 = fct_relevel(factor(X2), gsub("X", "", names(pw[,-1]))))

heat.l[[6]] <- ggplot() + 
  geom_tile(data=pw.long, aes(X, X2, fill=pFst)) +
  scale_fill_viridis_c(na.value = NA) +
  labs(title=ind.names[[6]], subtitle = paste0("Range is ", min(pw[pw>0], na.rm=T), "-", max(pw[,-1], na.rm = T)),
       y = "", x="") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 60, hjust=1)) 
ggsave(paste0(PATH, "/figures/pairwiseFst_heatmap_", ind.names[[6]], ".png"), plot = heat.l[[6]])

hc <- ggarrange(plotlist=heat.l[-6], ncol = 2, nrow = round(length(heat.l)/2))
ggsave(filename = paste0(PATH, "/figures/pairwiseFst_heatmap_comp.png"), plot = hc, width = 8, height = 8)
```
\begin{figure}
\centering
\includegraphics[width=0.9\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/pairwiseFst_heatmap_comp.png"}
\caption{Comparison of pairwise Fst values across groupings.}
\end{figure}
\pagebreak
\begin{figure}
\centering
\includegraphics[width=0.9\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/pairwiseFst_heatmap_Year_Pop.png"}
\caption{Temporal pairwise Fst. Looks like populations are more similar to themselves in other years than other populations pretty consistently.}
\end{figure}

```{r temp fst boxplot, message=F, eval=F}
#filter all rows that match for population
pw <- read.csv(paste0(PATH, "/results_tables/Year_Pop_pairwiseFst.csv")) %>%
  column_to_rownames("X")
pw[upper.tri(pw)] <- NA
pw <- pw %>% 
  rownames_to_column("X") %>%
  rename_with(.cols = starts_with("X2"), .fn = ~ gsub("X", "", .x)) %>%
  relocate('X', '2014_1', '2021_1',
           '2014_3', '2021_3',
           '2014_2', '2019_2', '2021_2',
           '2014_4', '2019_4', '2021_4',
           '2014_5', '2019_5', '2021_5',
           '2014_6', '2019_6', '2021_6',
           '2014_7', '2019_7', '2021_7',
           '2014_8', '2021_8',
           '2019_9', '2021_9')
pw.long <- pw %>%
  mutate(X = fct_relevel(factor(X), gsub("X", "", names(pw[,-1])))) %>%
  pivot_longer(cols = starts_with("20"), names_to = "X2", values_to = "pGst") %>%
  filter(!is.na(pGst)) %>%
  mutate(X2 = gsub("X", "", X2)) %>%
  mutate(X2 = fct_relevel(factor(X2), gsub("X", "", names(pw[,-1])))) %>%
  filter(X != X2)

pw.box <- pw.long %>% 
  mutate(pop1 = gsub("^[0-9]{4}_", "", X), 
         pop2 = gsub("^[0-9]{4}_", "", X2),
         year1 = as.numeric(gsub("_[0-9]*$", "", X)), 
         year2 = as.numeric(gsub("_[0-9]*$", "", X2))) #pull out pops + year
pw.box <- pw.box %>%
  mutate(group = case_when( 
    pop1 == pop2 & year1 - year2 > 0 ~ "Pop X vs Pop X\nin previous\nsample periods", #pop with itself in t-1 or t-2
    pop1 == pop2 & year2 - year1 > 0 ~ "Pop X vs Pop X\nin previous\nsample periods",
    pop1 != pop2 & year1 - year2 > 0 ~ "Pop X vs all other pops\nin previous\nsample periods", #pop with other pops in t-1 or t-2
    pop1 != pop2 & year2 - year1 > 0 ~ "Pop X vs all other pops\nin previous\nsample periods",
    pop1 != pop2 & year1 == year2 ~ "Pop X vs all other pops\nin the same\nsample period",
    TRUE ~ "other"
  ))
pw.box <- pw.box %>%
  mutate(group = fct_relevel(as.factor(group), c("Pop X vs Pop X\nin previous\nsample periods", "Pop X vs all other pops\nin previous\nsample periods", "Pop X vs all other pops\nin the same\nsample period")))

pfst.time <- ggplot(data = pw.box, aes(group, pGst)) +
  gghalves::geom_half_boxplot(aes(fill=group), errorbar.length = 0.25, show.legend = F, outlier.shape = NA) +
  gghalves::geom_half_point(color="black", alpha = 0.6) +
  scale_fill_manual(values = c("#F4D1AE","#12263A","#00C6B8")) +
  # geom_violin(draw_quantiles = TRUE, scale = "count") +
  labs(x="", y=bquote('pairwise' ~'F'[ST])) +
  theme(plot.background = element_rect(fill = "white"),
        panel.grid.major.y = element_line(color = "black", linetype = "dashed"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_rect(fill = "white"),
        axis.line = element_line(color = "black"),
        axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13))
pfst.time
ggsave(filename = paste0(PATH, "/figures/TemporalFst_Boxplot.png"), plot=pfst.time, width = 7, height = 6)
```
\begin{figure}
\centering
\includegraphics[width=0.9\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/TemporalFst_Boxplot.png"}
\caption{Temporal pairwise Fst - relationship between the same population at time t-1 or t-2 vs. each population against the others.}
\end{figure}
\pagebreak

---

\bigskip
### Proportion of shared alleles (Dps)  
```{r DPS, results='asis'}
for (i in 1:length(ind.list)) {
dps <- 1-pairwise.propShared(ind.list[[i]])
dps <- round(as.data.frame(as.matrix(dps)),3)
# dps[upper.tri(dps)] <- NA

print(kable(dps, caption=paste0("Dps for ", ind.names[[i]], ". Range is ", min(dps[dps > 0], na.rm = T), "-", max(dps, na.rm = T))))
write.csv(dps, file=paste0(PATH, "/results_tables/", ind.names[[i]], "_pairwiseDps.csv"))
}

pw <- as.data.frame(as.matrix(dps)) %>%
  relocate('2014_1', '2021_1',
           '2014_3', '2021_3',
           '2014_2', '2019_2', '2021_2',
           '2014_4', '2019_4', '2021_4',
           '2014_5', '2019_5', '2021_5',
           '2014_6', '2019_6', '2021_6',
           '2014_7', '2019_7', '2021_7',
           '2014_8', '2021_8',
           '2019_9', '2021_9')

pw.long <- pw %>%
  mutate(X = row.names(.)) %>%
  mutate(X = fct_relevel(factor(X), gsub("X", "", names(pw)))) %>%
  pivot_longer(cols = starts_with("20"), names_to = "X2", values_to = "Dps") %>%
  filter(!is.na(Dps)) %>%
  mutate(X2 = gsub("X", "", X2)) %>%
  mutate(X2 = fct_relevel(factor(X2), gsub("X", "", names(pw)))) %>%
  filter(X != X2)
  
write.csv(pw.long, file=paste0(PATH, "/results_tables/Year_Pop_Long_Dps.csv"), row.names = F)
```

```{r DPS figures, message=F, eval=F}
#heat map
heat.dps <- ggplot() + 
  geom_tile(data=pw.long, aes(X, X2, fill=Dps)) +
  scale_fill_viridis_c(na.value = NA) +
  labs(title=ind.names[[6]], subtitle = paste0("Range is ", min(pw[pw>0], na.rm=T), "-", max(pw[,-1], na.rm = T)),
       y = "", x="") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 60, hjust=1)) 
heat.dps
ggsave(paste0(PATH, "/figures/pairwiseDps_heatmap_", ind.names[[6]], ".png"), plot = heat.dps) 

#boxplot
pw <- read.csv(paste0(PATH, "/results_tables/Year_Pop_pairwiseDps.csv")) %>%
  column_to_rownames("X")
pw[upper.tri(pw)] <- NA
pw <- pw %>% 
  rownames_to_column("X") %>%
  rename_with(.cols = starts_with("X2"), .fn = ~ gsub("X", "", .x)) %>%
  relocate('X', '2014_1', '2021_1',
           '2014_3', '2021_3',
           '2014_2', '2019_2', '2021_2',
           '2014_4', '2019_4', '2021_4',
           '2014_5', '2019_5', '2021_5',
           '2014_6', '2019_6', '2021_6',
           '2014_7', '2019_7', '2021_7',
           '2014_8', '2021_8',
           '2019_9', '2021_9')
pw.long <- pw %>%
  mutate(X = fct_relevel(factor(X), gsub("X", "", names(pw[,-1])))) %>%
  pivot_longer(cols = starts_with("20"), names_to = "X2", values_to = "pGst") %>%
  filter(!is.na(pGst)) %>%
  mutate(X2 = gsub("X", "", X2)) %>%
  mutate(X2 = fct_relevel(factor(X2), gsub("X", "", names(pw[,-1])))) %>%
  filter(X != X2)

pw.box <- pw.long %>% 
  mutate(pop1 = gsub("^[0-9]{4}_", "", X), 
         pop2 = gsub("^[0-9]{4}_", "", X2),
         year1 = as.numeric(gsub("_[0-9]*$", "", X)), 
         year2 = as.numeric(gsub("_[0-9]*$", "", X2))) #pull out pops + year
pw.box <- pw.box %>%
  mutate(group = case_when( 
    pop1 == pop2 & year1 - year2 > 0 ~ "Pop X vs Pop X\nin previous\nsample periods", #pop with itself in t-1 or t-2
    pop1 == pop2 & year2 - year1 > 0 ~ "Pop X vs Pop X\nin previous\nsample periods",
    pop1 != pop2 & year1 - year2 > 0 ~ "Pop X vs all other pops\nin previous\nsample periods", #pop with other pops in t-1 or t-2
    pop1 != pop2 & year2 - year1 > 0 ~ "Pop X vs all other pops\nin previous\nsample periods",
    pop1 != pop2 & year1 == year2 ~ "Pop X vs all other pops\nin the same\nsample period",
    TRUE ~ "other"
  ))
pw.box <- pw.box %>%
  mutate(group = fct_relevel(as.factor(group), c("Pop X vs Pop X\nin previous\nsample periods", "Pop X vs all other pops\nin previous\nsample periods", "Pop X vs all other pops\nin the same\nsample period")))

dps.time <- ggplot(data = pw.box, aes(group, pGst)) +
  gghalves::geom_half_boxplot(aes(fill=group), errorbar.length = 0.25, show.legend = F, outlier.shape = NA) +
  gghalves::geom_half_point(color="black", alpha = 0.6) +
  scale_fill_manual(values = c("#F4D1AE","#12263A","#00C6B8")) +
  # geom_violin(draw_quantiles = TRUE, scale = "count") +
  labs(x="", y=bquote( ~'D'[PS])) +
  theme(plot.background = element_rect(fill = "white"),
        panel.grid.major.y = element_line(color = "black", linetype = "dashed"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_rect(fill = "white"),
        axis.line = element_line(color = "black"),
        axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13))
dps.time
ggsave(paste0(PATH, "/figures/TemporalDps_boxplot.png"), plot=dps.time, width = 7, height = 6)

pw.box %>% filter(group == "Pop A vs Pop A \nin previous years") %>% slice_max(Dps, n=5)
pw.box %>% filter(pop1 == "8" | pop2 == "8") %>% filter(pop1 != pop2) %>% mutate(mean = mean(Dps)) #avg. is 0.375
pw.box %>% filter(pop1 == "8" | pop2 == "8") %>% filter(pop1 == pop2) %>% mutate(mean = mean(Dps)) #0.308
```

### G~ST~  
*see Meirmans & Hedrick 2011*  

``` {r Gst global calc, message=F, eval=F}
library(mmod)

gst.list <- list()
for (i in 1:length(ind.list[-6])) {
  diff_stats(ind.list[[i]])
  b.g <- chao_bootstrap(ind.list[[i]], nreps = 10000)
  cat(ind.names[[i]], "complete...\n")
  b.gd <- summarise_bootstrap(b.g, Gst_Hedrick)
  gst.list[[i]] <- b.gd
}

global.gst <- as.data.frame(t(gst.list[[1]]$summary.global.het))
global.gst$group <- ind.names[[1]]
for (i in 2:length(gst.list)) {
  g <- as.data.frame(t(gst.list[[i]]$summary.global.het))
  g$group <- ind.names[[i]]
  global.gst <- bind_rows(global.gst, g)
}

global.gst <- global.gst %>% 
  mutate(group = case_when(group == "All Yrs - Indiv x Pop" ~ "Pop_Group",
                           group == "All Yrs - Indiv x Year" ~ "Year_Group",
                           group == "Indiv 2014" ~ "2014",
                           group == "Indiv 2019" ~ "2019",
                           group == "Indiv 2021" ~ "2021",
                           T ~ group)) %>%
  relocate(group)
write.csv(global.gst, file=paste0(PATH, "/results_tables/Gst_global.csv"), row.names = F)
```
``` {r gst dot plot, eval=F}
global.gst <- read.csv(paste0(PATH, "/results_tables/Gst_global.csv"))
global.gst <- global.gst %>% mutate(gr = case_when(group %in% c("Pop_Group", "Year_Group", "Year_Pop") ~ "all",
                                               group %in% c("2014", "2019", "2021") ~ "one"))

# global.gst <- read.csv(paste0(PATH, "/results_tables/Fst_global.csv"))
# global.gst <- global.gst %>% mutate(gr = case_when(group %in% c("All_Yrs", "All_Pops") ~ "all",
#                                                group %in% c("2014", "2019", "2021") ~ "one"),
#                                     observed = Fst, lower.normal = CI_ll, upper.normal = CI_ul)

pg <- ggplot(data = global.gst %>% filter(group != "Year_Pop"), aes(x=group, y=observed)) +
  geom_rect(xmin = 0, xmax = 3.5, ymin = -Inf, ymax = Inf, fill="#E5E4E2", alpha = 0.1) +
  geom_errorbar(aes(x=group, ymin=lower.normal, ymax=upper.normal), color="black", width=0.3) +
  geom_line(color="black") +
  geom_point(aes(fill=gr), shape=21, color="black", size=5.5) +
  scale_fill_manual(values = c("#12263A", "#F4D1AE")) +
  # scale_x_discrete(labels=c("2014", "2019", "2021", "All x\nPop", "All x\nYear", "All x\nYear+Pop")) +
  geom_vline(aes(xintercept=3.5), linetype="dashed") +
  labs(y=bquote('Global' ~'G"'[ST])) +
  # ggtitle(bquote(atop('Global' ~'G"'[ST]~ '','with 95% confidence intervals'))) +
  theme(legend.position = "none",
        panel.background = element_blank(),
        panel.grid = element_line(color="#E5E4E2"),
        title = element_text(size=17),
        plot.title = element_text(hjust=0.5),
        axis.title.x = element_blank(),
        axis.title.y = element_text(size = 15),
        axis.text.x = element_text(size = 15),
        axis.text.y = element_text(size = 12))
pg
ggsave(filename = paste0(PATH, "/figures/global_gst_points.png"), plot = pg, width = 6, height = 6)
```

``` {r gst heat and box, message=F, eval=F}
pw <- read.csv(paste0(PATH, "/results_tables/Year_Pop_pairwiseGst.csv")) %>% column_to_rownames("X")
  
pw[upper.tri(pw)] <- t(pw)[upper.tri(pw)]
pw <- pw %>% 
  rename_with(.cols = starts_with("X2"), .fn = ~ gsub("X", "", .x)) %>%
  relocate('2014_1', '2021_1',
           '2014_3', '2021_3',
           '2014_2', '2019_2', '2021_2',
           '2014_4', '2019_4', '2021_4',
           '2014_5', '2019_5', '2021_5',
           '2014_6', '2019_6', '2021_6',
           '2014_7', '2019_7', '2021_7',
           '2014_8', '2021_8',
           '2019_9', '2021_9')
pw.long <- pw %>%
  rownames_to_column("X") %>%
  mutate(X = fct_relevel(factor(X), gsub("X", "", names(pw)))) %>%
  pivot_longer(cols = starts_with("20"), names_to = "X2", values_to = "pGst") %>%
  filter(!is.na(pGst)) %>%
  mutate(X2 = gsub("X", "", X2)) %>%
  mutate(X2 = fct_relevel(factor(X2), gsub("X", "", names(pw)))) %>%
  filter(X != X2)
write.csv(pw.long, file=paste0(PATH, "/results_tables/Year_Pop_Long_pGst.csv"), row.names = F)

#heat map
heat.gst <- ggplot() + 
  geom_tile(data=pw.long, aes(X, X2, fill=pGst)) +
  scale_fill_viridis_c(na.value = NA) +
  labs(title=ind.names[[6]], subtitle = paste0("Range is ", min(pw[pw>0], na.rm=T), "-", max(pw[,-1], na.rm = T)),
       y = "", x="") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 60, hjust=1)) 
heat.gst
ggsave(paste0(PATH, "/figures/pairwiseGst_heatmap_", ind.names[[6]], ".png"), plot = heat.gst) 

#boxplot + summarize
pw <- read.csv(paste0(PATH, "/results_tables/Year_Pop_pairwiseGst.csv"))
pw <- pw %>% 
  rename_with(.cols = starts_with("X2"), .fn = ~ gsub("X", "", .x)) %>%
  relocate('X', '2014_1', '2021_1',
           '2014_3', '2021_3',
           '2014_2', '2019_2', '2021_2',
           '2014_4', '2019_4', '2021_4',
           '2014_5', '2019_5', '2021_5',
           '2014_6', '2019_6', '2021_6',
           '2014_7', '2019_7', '2021_7',
           '2014_8', '2021_8',
           '2019_9', '2021_9')
pw.long <- pw %>%
  mutate(X = fct_relevel(factor(X), gsub("X", "", names(pw[,-1])))) %>%
  pivot_longer(cols = starts_with("20"), names_to = "X2", values_to = "pGst") %>%
  filter(!is.na(pGst)) %>%
  mutate(X2 = gsub("X", "", X2)) %>%
  mutate(X2 = fct_relevel(factor(X2), gsub("X", "", names(pw[,-1])))) %>%
  filter(X != X2)

pw.box <- pw.long %>% 
  mutate(pop1 = gsub("^[0-9]{4}_", "", X), 
         pop2 = gsub("^[0-9]{4}_", "", X2),
         year1 = as.numeric(gsub("_[0-9]*$", "", X)), 
         year2 = as.numeric(gsub("_[0-9]*$", "", X2))) #pull out pops + year
pw.box <- pw.box %>%
  mutate(group = case_when( 
    pop1 == pop2 & year1 - year2 > 0 ~ "Pop X vs Pop X\nin previous\nsample periods", #pop with itself in t-1 or t-2
    pop1 != pop2 & year1 - year2 > 0 ~ "Pop X vs all other pops\nin previous\nsample periods", #pop with other pops in t-1 or t-2
    pop1 != pop2 & year1 == year2 ~ "Pop X vs all other pops\nin the same\nsample period",
    TRUE ~ "other"
  ))
pw.box <- pw.box %>%
  mutate(group = fct_relevel(as.factor(group), c("Pop X vs Pop X\nin previous\nsample periods", "Pop X vs all other pops\nin previous\nsample periods", "Pop X vs all other pops\nin the same\nsample period")))

gst.time <- ggplot(data = pw.box, aes(group, pGst)) +
  gghalves::geom_half_boxplot(aes(fill=group), errorbar.length = 0.25, show.legend = F, outlier.shape = NA) +
  gghalves::geom_half_point(color="black", alpha = 0.6) +
  scale_fill_manual(values = c("#F4D1AE","#12263A","#00C6B8")) +
  # geom_violin(draw_quantiles = TRUE, scale = "count") +
  labs(x="", y=bquote('pairwise' ~'G"'[ST])) +
  theme(plot.background = element_rect(fill = "white"),
        panel.grid.major.y = element_line(color = "black", linetype = "dashed"),
        panel.grid.major.x = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_rect(fill = "white"),
        axis.line = element_line(color = "black"),
        axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(size = 13))
gst.time
ggsave(paste0(PATH, "/figures/TemporalGst_boxplot.png"), plot=gst.time, width = 7, height = 6)

#summary of comparisons
pw.box %>% group_by(group) %>% summarise(mean = mean(pGst))
pw.box %>% filter(group == "Pop X vs Pop X\nin previous\nsample periods") %>% slice_max(pGst, n=5)
pw.box %>% filter(group != "Pop X vs Pop X\nin previous\nsample periods") %>% slice_max(pGst, n=10)
pw.box %>% filter(group != "Pop X vs Pop X\nin previous\nsample periods") %>% slice_min(pGst, n=10)
pw.box %>% filter(pop1 %in% c(3, 9) & pop2 %in% c(3, 9) & pop1 != pop2) %>% summarise(mean = mean(pGst))
pw.box %>% filter(pop1 %in% c(4, 5) & pop2 %in% c(4, 5) & pop1 != pop2) %>% summarise(mean = mean(pGst))
pw.box %>% filter(pop1 %in% c(5, 6) & pop2 %in% c(5, 6) & pop1 != pop2) %>% summarise(mean = mean(pGst))
```

\begin{figure}[!ht]
  \centering
  \includegraphics[width=0.5\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/global_gst_points.png"}
  \caption{Comparison of global Gst values with single year and multi-year groupings, population structure vs year structure.}
\end{figure}
<br>
<br>

---

\newpage

## Multivariate Analysis  
### PCA
*Note: Can only do PCA when matrix is Euclidean!*  
There are `r sum(is.na(ind.list[[1]]$tab))` missing data to be replaced  
Replaced missing data with mean allele frequency.  

``` {r pca set up, message=FALSE, eval=F}
library(gtools)
# replace missing values with mean of corresponding allele
x.list <- lapply(ind.list, function (x) gtools::na.replace(x$tab, mean, na.rm = TRUE) )

# x.list[[1]][1:5,1:5]
pca.list <- lapply(x.list, function(x) dudi.pca(x, scale = FALSE, scannf = FALSE, nf = 3))

# barplot(pca.list[[1]]$eig[1:50], main = paste0("PCA eigenvalues of ", ind.names[[1]]), col = heat.colors(50))
```

``` {r generic pca plots, fig.align="center", eval=F}
for (i in 2:length(pca.list)) {
  
  s.label(pca.list[[i]]$li) #li is the principle components, which in this case is summary of genetic diversity
  title(paste0("PCA of ", ind.names[[i]], " - axes 1-2"), line = 1)
  add.scatter.eig(pca.list[[i]]$eig[1:20], 2,1,2)
  
}

for (i in 1:length(pca.list)) {

    col <- virid(length(levels(ind.list[[i]]@pop)))
  s.class(pca.list[[i]]$li, pop(ind.list[[i]]),xax=1,yax=2, col=transp(col,.8), axesell=FALSE, cstar=0, cpoint=3, grid=FALSE)
  title(paste0("PCA of ", ind.names[[i]], " - axes 1-2"), line = 1)
  
}

for (i in 1:length(pca.list)) {

  s.class(pca.list[[i]]$li, pop(ind.list[[i]]), xax=1, yax=3, sub="PCA 1-3", csub=2)
  title(paste0("PCA of ", ind.names[[i]], " - axes 1-3"))
  add.scatter.eig(pca.list[[i]]$eig[1:20],nf=3,xax=1,yax=3)

}
```

``` {r pca ggplot, eval=F}
# library(ggforce)

pca.all <- pca.list[[1]]$li %>% 
  mutate(ID = row.names(.)) %>% 
  left_join(aztf, by = "ID") %>% 
  select(ID, year, pop, Axis1, Axis2) %>%
  mutate(pop = as.factor(as.numeric(pop)),
         year = fct_relevel(factor(year), c("2014", "2019", "2021")))

pca.theme <- list(
  scale_shape_manual(values = c(21, 22, 23)),
  theme(legend.position = "right", 
        panel.background = element_blank(), 
        panel.grid.major = element_line(colour = "gray", linetype = 4),
        axis.title = element_text(size = 23),
        axis.text = element_text(size = 20),
        legend.title = element_text(size = 20),
        legend.text = element_text(size = 18),
        legend.key = element_blank(),
        plot.title = element_text(face="bold",size=23)),
  scale_y_continuous(breaks = c(-3, 0, 3), limits = c(-3,4))
)

pca <- ggplot() +
  stat_ellipse(data = pca.all, aes(x=Axis1, y=Axis2, color=pop, group=pop), type = "t", alpha = 0.95, size=2, level = 0.9) +
  geom_point(data = pca.all, aes(x=Axis1, y=Axis2, fill=pop, shape=year), color="black", size=4.5, alpha=0.5) +
  labs(x="PC1 (6.16%)", y="PC2 (4.05%)", fill="Pond", color="Pond", shape="Year") +
  scale_fill_manual(values = aztf.pal, aesthetics = "fill") +
  scale_color_manual(values = aztf.pal, aesthetics = "color") +
  guides(fill = guide_legend(override.aes=list(shape=21))) +
  pca.theme
# pca

grays <- c("#CCCCCC", "#999999", "#333333")
pca2 <- ggplot() +
  stat_ellipse(data = pca.all, aes(x=Axis1, y=Axis2, color=year, group=year), type = "t", alpha = 0.95, size=2, level = 0.9) +
  geom_point(data = pca.all, aes(x=Axis1, y=Axis2, fill=year, shape=year), color="black", size=4.5, alpha=0.5) +
  labs(x="PC1 (6.16%)", y="PC2 (4.05%)", fill="Year", color="Year", shape="Year") +
  scale_fill_manual(values = grays, aesthetics = "fill") +
  scale_color_manual(values = grays, aesthetics = "color") +
  pca.theme
# pca2

ggsave(filename = paste0(PATH,"/figures/PCA_xYear.png"), plot = pca2, width = 8, height = 7)
ggsave(filename = paste0(PATH,"/figures/PCA_xPop.png"), plot = pca, width = 8, height = 7)
```

\begin{figure}[!hb]
  \centering
  \includegraphics[width=0.6\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/PCA_xYear.png"}
  \includegraphics[width=0.6\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/PCA_xPop.png"}
  \caption{PCA of all individuals across years grouped by either population or year. Shows higher structure spatially than temporally.}
\end{figure}

<br>
  
### DAPC  
Clusters were selected by lowest BIC, if more than one cluster had similar BIC, they all were looked at.  
When selecting the number of PCs, looked at PC x Cumulative Variance plot and visually picked, also used a score. Compared outputs from both.  
*DAPC run in 12d_DAPC.R script*  

\begin{figure}[!ht]
  \centering
  \includegraphics[width=0.9\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/DAPC_Indiv 2014_clustercomp.png"}
  \caption{DAPC of 2014 samples only and BIC values for iterations of K.}
\end{figure}

\begin{figure}[!ht]
  \centering
  \includegraphics[width=0.9\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/DAPC_Indiv 2019_clustercomp.png"}
  \caption{DAPC of 2019 samples only and BIC values for iterations of K.}
\end{figure}

\begin{figure}[!ht]
  \centering
  \includegraphics[width=0.9\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/DAPC_Indiv 2021_clustercomp.png"}
  \caption{DAPC of 2021 samples only and BIC values for iterations of K.}
\end{figure}

\begin{figure}[!ht]
  \centering
  \includegraphics[width=0.9\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/DAPC_All Yrs - Indiv x Pop_clustercomp.png"}
  \caption{DAPC of all years, with populations highlighted and BIC values for iterations of K.}
\end{figure}
