==> scripts/13_SpatialAnalysis.Rmd <==
---
title: "Spatial Analysis"
author: "C. E. Moore"
output: 
  pdf_document:
    latex_engine: pdflatex
urlcolor: blue
header-includes:
  \usepackage{helvet}
  \renewcommand{\familydefault}{\sfdefault}
---
Last updated on `r Sys.Date()`  
  
*Note: Microsat data previously formatted from geneious output using 01a_DataPrep.R and then filtered for missing data in 10_MicrosatSummary.Rmd*  

**Contents:**  
1. IBD w/ mantel  
  
```{r paths and data load, echo=FALSE, message=FALSE}
## packages
library(adegenet); library(knitr); library(tidyverse)

## home path
PATH <- "/home/chloe9mo/Documents/Projects/temporal_aztreefrog"

# gene data
source(file = paste0(PATH, "/code/01c_DataLoad.R"))

knitr::opts_chunk$set(echo = F, eval = F)
```
**Loaded packages:** `r (.packages())`  

# IBD with Mantel Test
``` {r IBD mantel}
source(paste0(PATH, "/code/12c_SpatialModels.R"))
```
``` {r IBD mantel tables, eval=T}
ibd_l <- lapply(list.files(path = paste0(PATH, "/results_tables"), pattern = "IBD_", full.names = T), read.csv)
kable(ibd_l[[1]], caption = "Dps mantel stats", digits = 3)
kable(ibd_l[[3]], caption = "Fst mantel stats", digits = 3)
kable(ibd_l[[2]], caption = "Fst mantel stats - log(geo.dist)", digits = 3)
kable(ibd_l[[4]], caption = "Gst mantel stats", digits = 3)
kable(ibd_l[[3]], caption = "Gst mantel stats - log(geo.dist)", digits = 3)
```
For each metric, almost all years have some IBD, with 2021 most significant, followed by 2019; none of the years follow IBstreamD though.

\begin{figure}[h!]
  \centering
  \includegraphics[width=0.4\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/IBD_Dps_2014.png"}
  \includegraphics[width=0.4\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/IBD_Dps_2019.png"}
  \includegraphics[width=0.4\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/IBD_Dps_2021.png"}
  \caption{IBD for three sample years, using proportion of shared alleles. Black line is lm(y~x), red line is loess smooth y~x.}
\end{figure}

\begin{figure}[!hb]
  \centering
  \includegraphics[width=0.4\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/Dist_x_Gst_2014 IBD.png"}
  \includegraphics[width=0.4\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/Dist_x_Gst_2019 IBD.png"}
  \includegraphics[width=0.4\textwidth]{"/home/chloe9mo/Documents/Projects/temporal_aztreefrog/figures/Dist_x_Gst_2021 IBD.png"}
  \caption{Individual year comparison of genetic x euclidian distance.}
\end{figure}

<br>
